# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Media Planner API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export type ContentType = (typeof ContentType)[keyof typeof ContentType];

export const ContentType = {
  video: "video",
  post: "post",
  article: "article",
} as const;

export type ContentStatus = (typeof ContentStatus)[keyof typeof ContentStatus];

export const ContentStatus = {
  draft: "draft",
  scheduled: "scheduled",
  published: "published",
} as const;

export type Platform = (typeof Platform)[keyof typeof Platform];

export const Platform = {
  instagram: "instagram",
  facebook: "facebook",
  twitter: "twitter",
  tiktok: "tiktok",
  youtube: "youtube",
  linkedin: "linkedin",
} as const;

export interface Content {
  id: number;
  title: string;
  caption: string;
  type: ContentType;
  status: ContentStatus;
  platforms: Platform[];
  /** ISO date string (YYYY-MM-DD) */
  scheduledDate: string;
  mediaUrl?: string | null;
  thumbnailUrl?: string | null;
  tags: string[];
  createdAt: string;
  updatedAt: string;
}

export interface CreateContentBody {
  title: string;
  caption: string;
  type: ContentType;
  status: ContentStatus;
  platforms: Platform[];
  /** ISO date string (YYYY-MM-DD) */
  scheduledDate: string;
  mediaUrl?: string | null;
  thumbnailUrl?: string | null;
  tags?: string[];
}

export interface UpdateContentBody {
  title?: string;
  caption?: string;
  type?: ContentType;
  status?: ContentStatus;
  platforms?: Platform[];
  /** ISO date string (YYYY-MM-DD) */
  scheduledDate?: string;
  mediaUrl?: string | null;
  thumbnailUrl?: string | null;
  tags?: string[];
}

export type DaySummaryByType = {
  video: number;
  post: number;
  article: number;
};

export type DaySummaryByStatus = {
  draft: number;
  scheduled: number;
  published: number;
};

export interface DaySummary {
  /** ISO date string (YYYY-MM-DD) */
  date: string;
  total: number;
  byType: DaySummaryByType;
  byStatus: DaySummaryByStatus;
}

export interface MonthSummary {
  month: number;
  year: number;
  days: DaySummary[];
  totalContent: number;
}

export type ContentStatsByType = {
  video: number;
  post: number;
  article: number;
};

export type ContentStatsByStatus = {
  draft: number;
  scheduled: number;
  published: number;
};

export type ContentStatsByPlatform = { [key: string]: number };

export interface ContentStats {
  total: number;
  byType: ContentStatsByType;
  byStatus: ContentStatsByStatus;
  byPlatform: ContentStatsByPlatform;
  publishedThisMonth: number;
  scheduledThisMonth: number;
}

export type ListContentParams = {
  /**
   * Filter by month (1-12)
   */
  month?: number;
  /**
   * Filter by year
   */
  year?: number;
  /**
   * Filter by content type
   */
  type?: ContentType;
  /**
   * Filter by status
   */
  status?: ContentStatus;
  /**
   * Filter by platform
   */
  platform?: string;
};

export type GetMonthSummaryParams = {
  /**
   * Month (1-12)
   */
  month: number;
  /**
   * Year
   */
  year: number;
};

export type GetContentStatsParams = {
  month?: number;
  year?: number;
};