# lib/api-zod/src/generated/api.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Media Planner API specification
 * OpenAPI spec version: 0.1.0
 */
import * as zod from "zod";

/**
 * Returns server health status
 * @summary Health check
 */
export const HealthCheckResponse = zod.object({
  status: zod.string(),
});

/**
 * @summary List all content items
 */
export const ListContentQueryParams = zod.object({
  month: zod.coerce.number().optional().describe("Filter by month (1-12)"),
  year: zod.coerce.number().optional().describe("Filter by year"),
  type: zod
    .enum(["video", "post", "article"])
    .optional()
    .describe("Filter by content type"),
  status: zod
    .enum(["draft", "scheduled", "published"])
    .optional()
    .describe("Filter by status"),
  platform: zod.coerce.string().optional().describe("Filter by platform"),
});

export const ListContentResponseItem = zod.object({
  id: zod.number(),
  title: zod.string(),
  caption: zod.string(),
  type: zod.enum(["video", "post", "article"]),
  status: zod.enum(["draft", "scheduled", "published"]),
  platforms: zod.array(
    zod.enum([
      "instagram",
      "facebook",
      "twitter",
      "tiktok",
      "youtube",
      "linkedin",
    ]),
  ),
  scheduledDate: zod.string().describe("ISO date string (YYYY-MM-DD)"),
  mediaUrl: zod.string().nullish(),
  thumbnailUrl: zod.string().nullish(),
  tags: zod.array(zod.string()),
  createdAt: zod.string(),
  updatedAt: zod.string(),
});
export const ListContentResponse = zod.array(ListContentResponseItem);

/**
 * @summary Create a new content item
 */
export const CreateContentBody = zod.object({
  title: zod.string(),
  caption: zod.string(),
  type: zod.enum(["video", "post", "article"]),
  status: zod.enum(["draft", "scheduled", "published"]),
  platforms: zod.array(
    zod.enum([
      "instagram",
      "facebook",
      "twitter",
      "tiktok",
      "youtube",
      "linkedin",
    ]),
  ),
  scheduledDate: zod.string().describe("ISO date string (YYYY-MM-DD)"),
  mediaUrl: zod.string().nullish(),
  thumbnailUrl: zod.string().nullish(),
  tags: zod.array(zod.string()).optional(),
});

/**
 * @summary Get a content item by ID
 */
export const GetContentParams = zod.object({
  id: zod.coerce.number(),
});

export const GetContentResponse = zod.object({
  id: zod.number(),
  title: zod.string(),
  caption: zod.string(),
  type: zod.enum(["video", "post", "article"]),
  status: zod.enum(["draft", "scheduled", "published"]),
  platforms: zod.array(
    zod.enum([
      "instagram",
      "facebook",
      "twitter",
      "tiktok",
      "youtube",
      "linkedin",
    ]),
  ),
  scheduledDate: zod.string().describe("ISO date string (YYYY-MM-DD)"),
  mediaUrl: zod.string().nullish(),
  thumbnailUrl: zod.string().nullish(),
  tags: zod.array(zod.string()),
  createdAt: zod.string(),
  updatedAt: zod.string(),
});

/**
 * @summary Update a content item
 */
export const UpdateContentParams = zod.object({
  id: zod.coerce.number(),
});

export const UpdateContentBody = zod.object({
  title: zod.string().optional(),
  caption: zod.string().optional(),
  type: zod.enum(["video", "post", "article"]).optional(),
  status: zod.enum(["draft", "scheduled", "published"]).optional(),
  platforms: zod
    .array(
      zod.enum([
        "instagram",
        "facebook",
        "twitter",
        "tiktok",
        "youtube",
        "linkedin",
      ]),
    )
    .optional(),
  scheduledDate: zod
    .string()
    .optional()
    .describe("ISO date string (YYYY-MM-DD)"),
  mediaUrl: zod.string().nullish(),
  thumbnailUrl: zod.string().nullish(),
  tags: zod.array(zod.string()).optional(),
});

export const UpdateContentResponse = zod.object({
  id: zod.number(),
  title: zod.string(),
  caption: zod.string(),
  type: zod.enum(["video", "post", "article"]),
  status: zod.enum(["draft", "scheduled", "published"]),
  platforms: zod.array(
    zod.enum([
      "instagram",
      "facebook",
      "twitter",
      "tiktok",
      "youtube",
      "linkedin",
    ]),
  ),
  scheduledDate: zod.string().describe("ISO date string (YYYY-MM-DD)"),
  mediaUrl: zod.string().nullish(),
  thumbnailUrl: zod.string().nullish(),
  tags: zod.array(zod.string()),
  createdAt: zod.string(),
  updatedAt: zod.string(),
});

/**
 * @summary Delete a content item
 */
export const DeleteContentParams = zod.object({
  id: zod.coerce.number(),
});

/**
 * @summary Get a summary of content for a given month
 */
export const GetMonthSummaryQueryParams = zod.object({
  month: zod.coerce.number().describe("Month (1-12)"),
  year: zod.coerce.number().describe("Year"),
});

export const GetMonthSummaryResponse = zod.object({
  month: zod.number(),
  year: zod.number(),
  days: zod.array(
    zod.object({
      date: zod.string().describe("ISO date string (YYYY-MM-DD)"),
      total: zod.number(),
      byType: zod.object({
        video: zod.number(),
        post: zod.number(),
        article: zod.number(),
      }),
      byStatus: zod.object({
        draft: zod.number(),
        scheduled: zod.number(),
        published: zod.number(),
      }),
    }),
  ),
  totalContent: zod.number(),
});

/**
 * @summary Get content statistics
 */
export const GetContentStatsQueryParams = zod.object({
  month: zod.coerce.number().optional(),
  year: zod.coerce.number().optional(),
});

export const GetContentStatsResponse = zod.object({
  total: zod.number(),
  byType: zod.object({
    video: zod.number(),
    post: zod.number(),
    article: zod.number(),
  }),
  byStatus: zod.object({
    draft: zod.number(),
    scheduled: zod.number(),
    published: zod.number(),
  }),
  byPlatform: zod.record(zod.string(), zod.number()),
  publishedThisMonth: zod.number(),
  scheduledThisMonth: zod.number(),
});