# lib/api-zod/src/generated/types/content.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Media Planner API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ContentStatus } from "./contentStatus";
import type { ContentType } from "./contentType";
import type { Platform } from "./platform";

export interface Content {
  id: number;
  title: string;
  caption: string;
  type: ContentType;
  status: ContentStatus;
  platforms: Platform[];
  /** ISO date string (YYYY-MM-DD) */
  scheduledDate: string;
  mediaUrl?: string | null;
  thumbnailUrl?: string | null;
  tags: string[];
  createdAt: string;
  updatedAt: string;
}