# lib/api-zod/src/generated/types/contentStats.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Media Planner API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ContentStatsByPlatform } from "./contentStatsByPlatform";
import type { ContentStatsByStatus } from "./contentStatsByStatus";
import type { ContentStatsByType } from "./contentStatsByType";

export interface ContentStats {
  total: number;
  byType: ContentStatsByType;
  byStatus: ContentStatsByStatus;
  byPlatform: ContentStatsByPlatform;
  publishedThisMonth: number;
  scheduledThisMonth: number;
}