# lib/api-zod/src/generated/types/contentStatus.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Media Planner API specification
 * OpenAPI spec version: 0.1.0
 */

export type ContentStatus = (typeof ContentStatus)[keyof typeof ContentStatus];

export const ContentStatus = {
  draft: "draft",
  scheduled: "scheduled",
  published: "published",
} as const;