# lib/api-zod/src/generated/types/contentType.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Media Planner API specification
 * OpenAPI spec version: 0.1.0
 */

export type ContentType = (typeof ContentType)[keyof typeof ContentType];

export const ContentType = {
  video: "video",
  post: "post",
  article: "article",
} as const;