# lib/api-zod/src/generated/types/daySummary.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Media Planner API specification
 * OpenAPI spec version: 0.1.0
 */
import type { DaySummaryByStatus } from "./daySummaryByStatus";
import type { DaySummaryByType } from "./daySummaryByType";

export interface DaySummary {
  /** ISO date string (YYYY-MM-DD) */
  date: string;
  total: number;
  byType: DaySummaryByType;
  byStatus: DaySummaryByStatus;
}