# lib/api-zod/src/generated/types/monthSummary.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Media Planner API specification
 * OpenAPI spec version: 0.1.0
 */
import type { DaySummary } from "./daySummary";

export interface MonthSummary {
  month: number;
  year: number;
  days: DaySummary[];
  totalContent: number;
}