# lib/api-zod/src/generated/types/platform.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Media Planner API specification
 * OpenAPI spec version: 0.1.0
 */

export type Platform = (typeof Platform)[keyof typeof Platform];

export const Platform = {
  instagram: "instagram",
  facebook: "facebook",
  twitter: "twitter",
  tiktok: "tiktok",
  youtube: "youtube",
  linkedin: "linkedin",
} as const;